+++ core/src/main/java/es/urjc/etsii/grafo/services/messaging/TelegramEventListener.java
@@ -4,13 +4,14 @@ import es.urjc.etsii.grafo.events.MorkEventListener;
 import es.urjc.etsii.grafo.events.types.ErrorEvent;
 import es.urjc.etsii.grafo.events.types.ExecutionEndedEvent;
 import es.urjc.etsii.grafo.events.types.ExperimentEndedEvent;
+import es.urjc.etsii.grafo.events.types.MorkEvent;
 
 import static es.urjc.etsii.grafo.util.TimeUtil.nanosToSecs;
 
 /**
  * Sends telegram messages on certain MorkEvents
  */
-public class TelegramEventListener {
+public class TelegramEventListener implements MorkEventListener {
 
     private final TelegramService telegramService;
     private volatile boolean errorNotified = false;
@@ -24,13 +25,23 @@ public class TelegramEventListener {
         this.telegramService = telegramService;
     }
 
+    @Override
+    public void onEvent(MorkEvent event) {
+        switch (event) {
+            case ExperimentEndedEvent experimentEndedEvent -> onExperimentEnd(experimentEndedEvent);
+            case ErrorEvent errorEvent -> onError(errorEvent);
+            case ExecutionEndedEvent ignored -> telegramService.stop();
+            default -> {
+            }
+        }
+    }
+
     /**
      * Send message when experiment ends
      *
      * @param event experiment ended event
      */
-    @MorkEventListener
-    public void onExperimentEnd(ExperimentEndedEvent event) {
+    private void onExperimentEnd(ExperimentEndedEvent event) {
         if (!telegramService.ready()) return;
         telegramService.sendMessage(String.format("Experiment %s ended. Execution time: %s seconds", event.experimentName(), nanosToSecs(event.executionTime())));
     }
@@ -40,8 +51,7 @@ public class TelegramEventListener {
      *
      * @param event error event
      */
-    @MorkEventListener
-    public void onError(ErrorEvent event) {
+    private void onError(ErrorEvent event) {
         if (!telegramService.ready()) return;
         // Only notify first error to prevent spamming
         if (!errorNotified) {
@@ -49,15 +59,5 @@ public class TelegramEventListener {
             var t = event.throwable();
             telegramService.sendMessage(String.format("Execution Error: %s. Further errors will NOT be notified.", t));
         }
-    }
-
-    /**
-     * Stop Telegram service on execution end
-     *
-     * @param event execution ended event
-     */
-    @MorkEventListener
-    public void onExecutionEnd(ExecutionEndedEvent event) {
-        telegramService.stop();
     }
 }